schedule-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
PyJWT==2.10.0
pydantic==2.8.2
python-dotenv==1.0.1